refactor(agents): split model types
Move model response, tool-call, usage, and metadata helper types into a sibling module while re-exporting them from moltis_agents::model. This keeps the public API stable and brings model.rs back under the repository file-size limit.
